+++ sdkoutput/Profile.js
@@ -8,15 +8,24 @@ export const Profile = {
    * Path: /profile
    * Responses: 201, 400, 500
    */
+  postProfile: async (token) => {
+    if (typeof token !== "string") {
+      throw new Error("Argument 'token' should be of type string");
+    }
 
-  postProfile: async (data = {}) => {
+    let config = {
+      url: `/profile`,
+      method: "post",
+      baseUrl: "https://example.com",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
+      },
+      data: {},
+    };
+
     try {
-      const response = await axios({
-        method: "post",
-        url: `/api/profile`,
-
-        data,
-      });
+      const response = await axios(config);
       return response.data;
     } catch (error) {
       console.error("Error in postProfile:", error);
@@ -31,14 +40,26 @@ export const Profile = {
    * Path: /profile
    * Responses: 200, 500
    */
+  getProfile: async (token, page) => {
+    if (typeof token !== "string") {
+      throw new Error("Argument 'token' should be of type string");
+    }
+    if (typeof page !== "integer") {
+      throw new Error("Argument 'page' should be of type integer ");
+    }
 
-  getProfile: async (params = {}) => {
+    let config = {
+      url: `/profile?page=${page}`,
+      method: "get",
+      baseUrl: "https://example.com",
+      headers: {
+        "Content-Type": "false",
+        Authorization: `Bearer ${token}`,
+      },
+    };
+
     try {
-      const response = await axios({
-        method: "get",
-        url: `/api/profile`,
-        params,
-      });
+      const response = await axios(config);
       return response.data;
     } catch (error) {
       console.error("Error in getProfile:", error);
@@ -53,14 +74,29 @@ export const Profile = {
    * Path: /profile/search
    * Responses: 200, 500
    */
+  getProfileSearch: async (token, username, email) => {
+    if (typeof token !== "string") {
+      throw new Error("Argument 'token' should be of type string");
+    }
+    if (typeof username !== "string") {
+      throw new Error("Argument 'username' should be of type string ");
+    }
+    if (typeof email !== "string") {
+      throw new Error("Argument 'email' should be of type string ");
+    }
 
-  getProfileSearch: async (params = {}) => {
+    let config = {
+      url: `/profile/search?username=${username}&email=${email}`,
+      method: "get",
+      baseUrl: "https://example.com",
+      headers: {
+        "Content-Type": "false",
+        Authorization: `Bearer ${token}`,
+      },
+    };
+
     try {
-      const response = await axios({
-        method: "get",
-        url: `/api/profile/search`,
-        params,
-      });
+      const response = await axios(config);
       return response.data;
     } catch (error) {
       console.error("Error in getProfileSearch:", error);
@@ -75,14 +111,26 @@ export const Profile = {
    * Path: /profile/{userId}
    * Responses: 200, 404, 500
    */
+  getProfileByUserId: async (token, userId) => {
+    if (typeof token !== "string") {
+      throw new Error("Argument 'token' should be of type string");
+    }
+    if (typeof userId !== "string") {
+      throw new Error("Argument 'userId' should be of type string ");
+    }
 
-  getProfileByUserId: async (params = {}) => {
+    let config = {
+      url: `/profile/${userId}`,
+      method: "get",
+      baseUrl: "https://example.com",
+      headers: {
+        "Content-Type": "false",
+        Authorization: `Bearer ${token}`,
+      },
+    };
+
     try {
-      const response = await axios({
-        method: "get",
-        url: `/api/profile/{userId}`,
-        params,
-      });
+      const response = await axios(config);
       return response.data;
     } catch (error) {
       console.error("Error in getProfileByUserId:", error);
@@ -97,15 +145,40 @@ export const Profile = {
    * Path: /profile/{userId}
    * Responses: 200, 400, 404, 500
    */
+  putProfileByUserId: async (token, userId, username, email, profileImage) => {
+    if (typeof token !== "string") {
+      throw new Error("Argument 'token' should be of type string");
+    }
+    if (typeof userId !== "string") {
+      throw new Error("Argument 'userId' should be of type string ");
+    }
+    if (typeof username !== "string") {
+      throw new Error("Argument 'username' should be of type string ");
+    }
+    if (typeof email !== "string") {
+      throw new Error("Argument 'email' should be of type string ");
+    }
+    if (typeof profileImage !== "string") {
+      throw new Error("Argument 'profileImage' should be of type string ");
+    }
 
-  putProfileByUserId: async (params = {}, data = {}) => {
+    let config = {
+      url: `/profile/${userId}`,
+      method: "put",
+      baseUrl: "https://example.com",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
+      },
+      data: {
+        username: `${username}`,
+        email: `${email}`,
+        profileImage: `${profileImage}`,
+      },
+    };
+
     try {
-      const response = await axios({
-        method: "put",
-        url: `/api/profile/{userId}`,
-        params,
-        data,
-      });
+      const response = await axios(config);
       return response.data;
     } catch (error) {
       console.error("Error in putProfileByUserId:", error);
@@ -120,14 +193,26 @@ export const Profile = {
    * Path: /profile/{userId}
    * Responses: 200, 404, 500
    */
+  deleteProfileByUserId: async (token, userId) => {
+    if (typeof token !== "string") {
+      throw new Error("Argument 'token' should be of type string");
+    }
+    if (typeof userId !== "string") {
+      throw new Error("Argument 'userId' should be of type string ");
+    }
 
-  deleteProfileByUserId: async (params = {}) => {
+    let config = {
+      url: `/profile/${userId}`,
+      method: "delete",
+      baseUrl: "https://example.com",
+      headers: {
+        "Content-Type": "false",
+        Authorization: `Bearer ${token}`,
+      },
+    };
+
     try {
-      const response = await axios({
-        method: "delete",
-        url: `/api/profile/{userId}`,
-        params,
-      });
+      const response = await axios(config);
       return response.data;
     } catch (error) {
       console.error("Error in deleteProfileByUserId:", error);
@@ -142,15 +227,36 @@ export const Profile = {
    * Path: /profile/{userId}/upload
    * Responses: 200, 400, 500
    */
+  postProfileByUserIdUpload: async (token, userId, files) => {
+    if (typeof token !== "string") {
+      throw new Error("Argument 'token' should be of type string");
+    }
+    if (typeof userId !== "string") {
+      throw new Error("Argument 'userId' should be of type string ");
+    }
+    if (!Array.isArray(files)) {
+      throw new Error(
+        "Argument 'undefined' should be of type array and the items should be of the type in the {'type':'array','items':{'type':'string','format':'binary'}} format",
+      );
+    }
 
-  postProfileByUserIdUpload: async (params = {}, data = {}) => {
+    let formData = new FormData();
+
+    formData.append("files", files);
+
+    let config = {
+      url: `/profile/${userId}/upload`,
+      method: "post",
+      baseUrl: "https://example.com",
+      headers: {
+        "Content-Type": "multipart/form-data",
+        Authorization: `Bearer ${token}`,
+      },
+      data: formData,
+    };
+
     try {
-      const response = await axios({
-        method: "post",
-        url: `/api/profile/{userId}/upload`,
-        params,
-        data,
-      });
+      const response = await axios(config);
       return response.data;
     } catch (error) {
       console.error("Error in postProfileByUserIdUpload:", error);
